Poprawki lista 1

+++ lista1/zadanie3_lista1.py
@@ -1,8 +1,8 @@
 import math
 
-a = int(input("Podaj a: "))
-b = int(input("Podaj b: "))
-alfa = int(input("Podaj kąt: "))
+a = float(input("Podaj a: "))
+b = float(input("Podaj b: "))
+alfa = float(input("Podaj kąt: "))
 print("Pole wynosi: ")
 pole = 1 / 2 * a * b * math.sin(math.radians(alfa))
 print(pole)
